Ensure users can only delete their own jokes
Relates #44

+++ src/routes/delete.js
@@ -1,9 +1,12 @@
 const delete_joke = require('./../delete_joke.js');
 
 module.exports = {
-  path: '/delete/{id}',
+  path: '/delete/{username}/{id}',
   method: 'GET',
   handler: (req, reply) => {
+    if (req.auth.credentials.username !== req.params.username) {
+      return reply.redirect('/');
+    }
     delete_joke(req.params.id, (err) => {
       if (err) {
         return reply('Something went wrong, sorry!');
